v0.0.5 & Fixed integer being converted to integer (it's supposed to be long, thanks to Discord)

+++ src/main/java/dev/robothanzo/jda/interactions/JDAInteractions.java
@@ -51,7 +51,7 @@ import java.util.Objects;
 public class JDAInteractions {
     public static final Map<OptionType, Class<?>> OPTION_TYPE_CLASS_MAP = Map.of(
             OptionType.STRING, String.class,
-            OptionType.INTEGER, Integer.class,
+            OptionType.INTEGER, Long.class,
             OptionType.NUMBER, Double.class,
             OptionType.BOOLEAN, Boolean.class,
             OptionType.USER, User.class,
